Added Trigger Tank Drive
This drive mode is similar to tank drive, but when a trigger is held, the drive mode switches. This new drive mode reduces the maximum output of the joysticks to 25%, but holding down one trigger adds up to 75% output power, and holding down the other adds 75% output power in the opposite direction.